requirements.txt
pandas
openpyxl
langchain
langgraph
streamlit
google-auth
python-dotenv
google-generativeai
langchain-google-genai
